cleanup .net build changes
