+++ Assets/MixedRealityToolkit/_Core/Attributes/ExtendsAttribute.cs
@@ -3,6 +3,7 @@
 
 using System;
 using Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities;
+using Microsoft.MixedReality.Toolkit.Internal.Extensions;
 
 namespace Microsoft.MixedReality.Toolkit.Internal.Attributes
 {
