+++ Assets/MixedRealityToolkit/_Core/Attributes/ImplementsAttribute.cs
@@ -1,8 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using System;
 using Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities;
+using Microsoft.MixedReality.Toolkit.Internal.Extensions;
+using System;
 
 namespace Microsoft.MixedReality.Toolkit.Internal.Attributes
 {
