+++ Assets/MixedRealityToolkit/_Core/Definitions/Utilities/SystemType.cs
@@ -1,6 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+#if WINDOWS_UWP && !ENABLE_IL2CPP
+using Microsoft.MixedReality.Toolkit.Internal.Extensions;
+#endif // WINDOWS_UWP && !ENABLE_IL2CPP
 using System;
 using UnityEngine;
 
@@ -24,12 +27,9 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
                 return string.Empty;
             }
 
-#if WINDOWS_UWP && !ENABLE_IL2CPP
-            return $"{type.FullName}, {type.GetAssembly().GetName().Name}";
-#else
-            return $"{type.FullName}, {type.Assembly.GetName().Name}";
-#endif
-
+            string[] qualifiedNameComponents = type.AssemblyQualifiedName.Split(new char[] { ',' });
+            Debug.Assert(qualifiedNameComponents.Length >= 2);
+            return $"{qualifiedNameComponents[0]}, {qualifiedNameComponents[1].Trim()}";
         }
 
         /// <summary>
@@ -55,7 +55,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
             Type = type;
         }
 
-        #region ISerializationCallbackReceiver Members
+#region ISerializationCallbackReceiver Members
 
         void ISerializationCallbackReceiver.OnAfterDeserialize()
         {
@@ -76,7 +76,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
 
         void ISerializationCallbackReceiver.OnBeforeSerialize() { }
 
-        #endregion ISerializationCallbackReceiver Members
+#endregion ISerializationCallbackReceiver Members
 
         /// <summary>
         /// Gets or sets type of class reference.
@@ -93,7 +93,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
                 bool isValid = value.IsValueType() && !value.IsEnum() || value.IsClass();
 #else
                 bool isValid = value.IsValueType && !value.IsEnum || value.IsClass;
-#endif
+#endif // WINDOWS_UWP && !ENABLE_IL2CPP
                 if (value != null && !isValid)
                 {
                     throw new ArgumentException($"'{value.FullName}' is not a class or struct type.", nameof(value));
