--- Assets/MixedRealityToolkit/_Core/Extensions/NetNativeReflectionExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b256470477a2c53479a8fbddbfc6eedd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8a1b00ef1a7a8b94891466c1f0911027, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
